refactor(register): 상품 정보 input 구조 리팩토링
