+++ app/(routes)/artists/products/register/_components/PriceInputGroup.tsx
