+++ app/(routes)/artists/products/register/_components/ProductInfoSection.tsx
